+++ lib/core.dart
@@ -1,5 +1,6 @@
 library core;
 
+export 'src/core/console.dart';
 export 'src/core/hash_codes.dart';
 export 'src/core/timestamp.dart';
 export 'src/core/utils.dart';