+++ lib/src/core/console_stub.dart
@@ -1,0 +1,4 @@
+import 'console.dart';
+
+Console createConsole() => throw UnsupportedError(
+    "Can't create a console without dart:html or dart:io.");