+++ lib/src/signalr/server_sent_events_transport.dart
@@ -59,7 +59,7 @@ class ServerSentEventsTransport implements Transport {
     }
 
     EventSource eventSource;
-    if (Platform.isWeb) {
+    if (Platform.isChromium) {
       eventSource = _eventSourceConstructor(url);
     } else {
       final headers = <String, String>{};