+++ lib/src/signalr/utils_chromium.dart
@@ -1,15 +1,8 @@
-import 'dart:html';
-
 String get osName => '';
 String get runtime => 'Browser';
 String get runtimeVersion => null;
 
 String get userAgent => 'X-SignalR-User-Agent';
 
-bool get isWeb => true;
-bool get isVM => false;
-
-void error(Object message) => window.console.error(message);
-void warn(Object message) => window.console.warn(message);
-void info(Object message) => window.console.info(message);
-void log(Object message) => window.console.log(message);
+bool get isChromium => true;
+bool get isDartium => false;