+++ lib/src/signalr/utils_dartium.dart
@@ -6,10 +6,5 @@ String get runtimeVersion => Platform.version;
 
 String get userAgent => 'User-Agent';
 
-bool get isWeb => false;
-bool get isVM => true;
-
-void error(Object message) => print(message);
-void warn(Object message) => print(message);
-void info(Object message) => print(message);
-void log(Object message) => print(message);
+bool get isChromium => false;
+bool get isDartium => true;