+++ lib/src/signalr/utils_stub.dart
@@ -4,10 +4,5 @@ String get runtimeVersion => throw UnimplementedError();
 
 String get userAgent => throw UnimplementedError();
 
-bool get isWeb => throw UnimplementedError();
-bool get isVM => throw UnimplementedError();
-
-void error(Object message) => throw UnimplementedError();
-void warn(Object message) => throw UnimplementedError();
-void info(Object message) => throw UnimplementedError();
-void log(Object message) => throw UnimplementedError();
+bool get isChromium => throw UnimplementedError();
+bool get isDartium => throw UnimplementedError();